Update and rename custom.tsx to SteeringWheel.tsx
