+++ components/icon/demo/SteeringWheel.tsx
@@ -1,0 +1,32 @@
+import React from 'react';
+import Icon, { HomeOutlined } from '@ant-design/icons';
+import { Space } from 'antd';
+import type { GetProps } from 'antd';
+
+type CustomIconComponentProps = GetProps<typeof Icon>;
+
+const HeartSvg = () => (
+  <svg width="1em" height="1em" fill="currentColor" viewBox="0 0 1024 1024">
+    <title>heart icon</title>
+    <path d="M923 283.6c-13.4-31.1-32.6-58.9-56.9-82.8-24.3-23.8-52.5-42.4-84-55.5-32.5-13.5-66.9-20.3-102.4-20.3-49.3 0-97.4 13.5-139.2 39-10 6.1-19.5 12.8-28.5 20.1-9-7.3-18.5-14-28.5-20.1-41.8-25.5-89.9-39-139.2-39-35.5 0-69.9 6.8-102.4 20.3-31.4 13-59.7 31.7-84 55.5-24.4 23.9-43.5 51.7-56.9 82.8-13.9 32.3-21 66.6-21 101.9 0 33.3 6.8 68 20.3 103.3 11.3 29.5 27.5 60.1 48.2 91 32.8 48.9 77.9 99.9 133.9 151.6 92.8 85.7 184.7 144.9 188.6 147.3l23.7 15.2c10.5 6.7 24 6.7 34.5 0l23.7-15.2c3.9-2.5 95.7-61.6 188.6-147.3 56-51.7 101.1-102.7 133.9-151.6 20.7-30.9 37-61.5 48.2-91 13.5-35.3 20.3-70 20.3-103.3 0.1-35.3-7-69.6-20.9-101.9z" />
+  </svg>
+);
+
+const SteeringWheelSvg = () => (
+<svg fill="#ffffff" version="1.1" id="Capa_1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" viewBox="0 0 402.979 402.98" xml:space="preserve"><g id="SVGRepo_bgCarrier" stroke-width="0"></g><g id="SVGRepo_tracerCarrier" stroke-linecap="round" stroke-linejoin="round"></g><g id="SVGRepo_iconCarrier"> <g> <path d="M401.098,173.863c-3.679-26.537-12.961-52.711-26.852-75.692c-13.571-22.455-31.934-42.557-53.098-58.134 c-21.481-15.808-46.475-27.396-72.278-33.514c-15.063-3.57-30.615-5.381-46.225-5.381c-0.385,0-0.771,0.013-1.156,0.016 c-0.385-0.003-0.77-0.016-1.156-0.016c-15.607,0-31.16,1.811-46.225,5.381c-25.805,6.117-50.799,17.706-72.277,33.514 C60.667,55.614,42.306,75.716,28.734,98.171c-13.89,22.981-23.175,49.155-26.852,75.692c-3.804,27.466-1.819,56.001,5.745,82.523 c7.252,25.415,19.898,49.799,36.568,70.516c16.494,20.496,37.385,37.979,60.41,50.555c23.664,12.924,50.357,21.062,77.193,23.533 c6.107,0.562,12.32,0.848,18.467,0.848h2.445c6.148,0,12.361-0.285,18.47-0.848c26.836-2.473,53.526-10.609,77.19-23.533 c23.026-12.576,43.914-30.059,60.411-50.555c16.671-20.717,29.315-45.101,36.567-70.516 C402.915,229.865,404.901,201.33,401.098,173.863z M166.992,355.189c-26.398-5.723-51.072-18.324-71.354-36.449 c-20.141-18.004-35.394-40.938-44.107-66.326c-10.244-29.846-11.169-63.041-2.609-93.468c8.938,7.768,18.547,16.027,28.264,23.806 c7.611,6.092,15.258,11.93,22.729,17.354c3.699,2.685,7.578,5.405,11.525,8.085c1.929,1.307,3.863,2.59,5.756,3.815 c0.861,0.56,1.73,1.116,2.594,1.666c0.984,0.624,1.234,0.923,1.681,2.005c7.584,18.4,22.036,31.432,32.827,39.123 c3.66,2.611,7.479,4.988,11.348,7.062c1.068,0.573,2.148,1.129,3.24,1.662c0.084,1.008,0.172,2.014,0.268,3.012 c0.205,2.177,0.422,4.347,0.662,6.518c0.982,9.012,2.264,18.322,3.918,28.462c2.938,18.015,6.678,36.497,11.434,56.505 C185.166,358.021,179.59,357.769,166.992,355.189z M198.935,145.043c-13.793,0-25.855,1.274-36.873,3.897 c-3.471,0.827-6.793,1.806-9.873,2.908c-1.592,0.568-3.164,1.188-4.67,1.841c-0.685,0.295-1.366,0.603-2.041,0.92 c-0.719,0.336-1.363,0.065-1.999-0.254c-0.163-0.084-0.341-0.172-0.528-0.254c-7.701-3.424-16.213-6.88-26.01-10.561 c-17.166-6.452-35.454-12.459-54.354-17.856v-0.001c12.685-22.957,31.079-42.576,53.197-56.737 c25.254-16.171,54.492-24.718,84.549-24.718h2.313c30.057,0,59.295,8.547,84.549,24.718c22.118,14.161,40.514,33.78,53.197,56.737 v0.001c-18.898,5.397-37.188,11.404-54.354,17.856c-9.797,3.681-18.307,7.137-26.01,10.561c-0.188,0.082-0.363,0.17-0.527,0.254 c-0.638,0.319-1.28,0.59-2,0.254c-0.674-0.317-1.354-0.625-2.039-0.92c-1.508-0.652-3.079-1.272-4.672-1.841 c-3.078-1.103-6.399-2.081-9.872-2.908c-11.018-2.623-23.08-3.897-36.872-3.897H198.935z M351.449,252.414 c-8.713,25.389-23.967,48.322-44.107,66.326c-20.279,18.125-44.954,30.729-71.352,36.449c-11.934,2.469-18.176,2.832-18.176,2.832 c4.754-20.009,8.496-38.49,11.435-56.506c1.653-10.139,2.936-19.449,3.92-28.462c0.237-2.171,0.454-4.341,0.66-6.517 c0.096-1,0.184-2.004,0.268-3.012c1.094-0.535,2.17-1.09,3.24-1.662c3.868-2.076,7.688-4.453,11.348-7.063 c10.791-7.69,25.241-20.722,32.828-39.123c0.444-1.082,0.694-1.381,1.68-2.005c0.863-0.549,1.731-1.105,2.596-1.666 c1.891-1.227,3.826-2.51,5.754-3.816c3.947-2.68,7.826-5.398,11.525-8.084c7.471-5.424,15.119-11.262,22.727-17.354 c9.722-7.778,19.329-16.038,28.267-23.806C362.619,189.373,361.693,222.568,351.449,252.414z"></path> </g> </g></svg>
+);
+
+const SteeringWheel = (props: Partial<CustomIconComponentProps>) => (
+  <Icon component={SteeringWheelSvg} {...props} />
+);
+
+const App: React.FC = () => (
+  <Space>
+    <HeartIcon style={{ color: 'hotpink' }} />
+    <SteeringWheel style={{ fontSize: '32px' }} />
+    <Icon component={HomeOutlined as React.ForwardRefExoticComponent<any>} />
+    <HomeOutlined />
+  </Space>
+);
+
+export default App;
